scripts/bot/requirements.txt
pyTelegramBotAPI
psutil
python-dotenv
requests
qrcode
Pillow
